+++ package.json
@@ -2,13 +2,13 @@
   "name": "completeline",
   "displayName": "Complete Line",
   "description": "Completes a line of code",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "publisher": "refactor11",
   "author": {
     "name": "Jakob Christensen"
   },
   "engines": {
-    "vscode": "^1.56.0"
+    "vscode": "^1.105.0"
   },
   "categories": [
     "Other"
@@ -47,7 +47,7 @@
     "vscode:prepublish": "npm run compile",
     "compile": "tsc -p ./",
     "watch": "tsc -watch -p ./",
-    "package": "vsce package",
+    "package": "npx vsce package",
     "postinstall": "node ./node_modules/vscode/bin/install",
     "test": "npm run compile && node ./node_modules/vscode/bin/test"
   },
