+++ src/constants.ts
@@ -1,0 +1,34 @@
+export const BLOCK_KEYWORDS = [
+  "if",
+  "for",
+  "foreach",
+  "while",
+  "do",
+  "function",
+  "else",
+  "class",
+  "switch",
+  "try",
+  "catch",
+  "with",
+  "public",
+  "protected",
+  "private",
+  "internal",
+  "void",
+];
+
+export const LOOP_KEYWORDS = ["for", "foreach", "do", "while"];
+
+export const SUPPORTED_LANGUAGES = [
+  "typescript",
+  "typescriptreact",
+  "javascript",
+  "javascriptreact",
+  "csharp",
+  "c",
+  "cpp",
+  "rust",
+  "vue",
+  "java",
+];
