+++ src/types/SnippetQuickPickItem.ts
@@ -1,0 +1,5 @@
+import * as vscode from "vscode";
+
+export interface SnippetQuickPickItem extends vscode.QuickPickItem {
+  snippet: vscode.SnippetString;
+}
